requirements.txt
GPUtil==1.4.0
dataclasses==0.6
hydra_core==1.3.2
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
prettytable==3.9.0
py-cpuinfo==9.0.0
scikit-learn==1.3.1
scipy==1.10.1
tensorflow-probability==0.21.0
tensorflow==2.13.1
tqdm==4.66.1
xgboost==2.0.0
tabulate==0.9.0
xlsxwriter==3.2.0
pre-commit
